some cleanup and restructuring
