+++ pkg/reconciler/matcher_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rbacdefinition
+package reconciler
 
 import (
 	"testing"
